requirements.txt
# Core
python-dotenv==1.0.0
APScheduler==3.10.4

# YouTube
yt-dlp
google-api-python-client==2.108.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# IA
google-generativeai==0.3.1

# TTS
edge-tts==6.1.9

# Database
sqlalchemy>=2.0.35

# Utils
requests==2.31.0
python-dateutil==2.8.2
isodate==0.6.1
